wait for leader election before submitting the consensus operation
