+++ .dockerignore
@@ -6,4 +6,5 @@ lib/segment/target
 lib/api/target
 lib/collection/target
 lib/storage/target
+openapi/tests/
 *.tar
