+++ lib/storage/src/content_manager/consensus/mod.rs
@@ -1,4 +1,5 @@
 pub mod consensus_wal;
 pub mod entry_queue;
+pub mod is_ready;
 pub mod operation_sender;
 pub mod persistent;
