+++ src/common/helpers.rs
@@ -1,6 +1,5 @@
 use std::sync::atomic::{AtomicUsize, Ordering};
 
-use parking_lot::{Condvar, Mutex};
 use tokio::runtime;
 use tokio::runtime::Runtime;
 
@@ -22,45 +21,6 @@ pub fn create_search_runtime(max_search_threads: usize) -> std::io::Result<Runti
         .build()
 }
 
-pub struct IsReady {
-    condvar: Condvar,
-    value: Mutex<bool>,
-}
-
-impl Default for IsReady {
-    fn default() -> Self {
-        Self {
-            condvar: Condvar::new(),
-            value: Mutex::new(false),
-        }
-    }
-}
-
-impl IsReady {
-    pub fn make_ready(&self) {
-        let mut is_ready = self.value.lock();
-        if !*is_ready {
-            *is_ready = true;
-            self.condvar.notify_all();
-        }
-    }
-
-    pub fn make_not_ready(&self) {
-        *self.value.lock() = false;
-    }
-
-    pub fn check_ready(&self) -> bool {
-        *self.value.lock()
-    }
-
-    pub fn await_ready(&self) {
-        let mut is_ready = self.value.lock();
-        if !*is_ready {
-            self.condvar.wait(&mut is_ready);
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
@@ -68,7 +28,7 @@ mod tests {
     use std::thread::sleep;
     use std::time::Duration;
 
-    use super::*;
+    use storage::content_manager::consensus::is_ready::IsReady;
 
     #[test]
     fn test_is_ready() {
